update format perferred name
